redroid-manager/backend/requirements.txt
fastapi
uvicorn
docker
pydantic
sqlalchemy
passlib[bcrypt]
pyjwt
python-multipart
